% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{getCQL}
\alias{getCQL}
\title{Query CQL}
\usage{
getCQL(
  cqlArr = NULL,
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL
)
}
\arguments{
\item{corpusName}{Name of corpus to query}

\item{corpora}{Name of corpora(s) to query}

\item{lang}{Query by Language}

\item{media}{Query by Media}

\item{age}{Query by Age}

\item{gender}{Query by gender}

\item{designType}{Query by design type}

\item{activityType}{Query by activity type}

\item{groupType}{Query by group type}
}
\description{
The "CQL" (Corpus Query Language) tab lets us search for patterns in the selected transcripts. We construct a CQL query by appending words, lemmas, and parts of speech. Under 'Insert' in the CQL tab we assemble a query by specifying and inserting a word, lemma, or part of speech. Inserting a 'Word' matches that exact word. Inserting a part of speech matches a word with that part of speech. Inserting a 'Lemma' matches that word stem. For instance, inserting lemma 'run' matches:
\itemize{
\item run
\item running
\item ran
}
}
\examples{
cqlMyBall <- getCQL(cqlArr=list(list(type="lemma", item="my", freq="once"), list(type="lemma", item="ball", freq="once")), corpusName = list('childes'), corpora = list(list('childes', 'Eng-NA', 'MacWhinney')))
cqlEat <- getCQL(cqlArr=list(list(type="lemma", item="eat", freq="once")), corpusName = list('childes'), corpora = list(list('childes', 'Eng-NA', 'MacWhinney')))
}
