% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getLegalValues}
\alias{getLegalValues}
\title{Get legal values}
\usage{
getLegalValues()
}
\description{
Prints available options (values) for each level of talkbankDB
\itemize{
\item After running getLegalValues(), select the corpus you would like to query.
\item The function will return the next level of options
\item Choose from those options to progres to the next level of available options
\item All names are case-sensitive
}
}
\examples{
getLegalValues()
}
