% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{getParticipants}
\alias{getParticipants}
\title{Get participant data}
\usage{
getParticipants(
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL
)
}
\arguments{
\item{corpusName}{Name of corpus to query}

\item{corpora}{Name of corpora(s) to query}

\item{lang}{Query by Language}

\item{media}{Query by Media}

\item{age}{Query by Age}

\item{gender}{Query by gender}

\item{designType}{Query by design type}

\item{activityType}{Query by activity type}

\item{groupType}{Query by group type}
}
\description{
Queries a table where each row represents a participant (speaker) listed in a transcript. Each row has:
\itemize{
\item Link to view transcript and play any associated media.
\item Corpus path to transcript.
\item Speaker's ID.
\item Speaker's name.
\item Speaker's role.
\item Speaker's language.
\item Speaker's age in months.
\item Speaker's age in Years/Months/Days.
\item Speaker's gender.
\item Number of words spoken by speaker.
\item Number of utterances spoken by speaker.
\item Average number of words per speaker's utterance.
\item Median number of words per speaker's utterance.
}
}
\examples{
getParticipants(corpusName = list('childes'),
                corpora = list(list('childes',
                                    'Eng-NA',
                                    'MacWhinney')))
}
