% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getPathTrees}
\alias{getPathTrees}
\title{Get Path Trees}
\usage{
getPathTrees()
}
\description{
Get path tree to every doc in TalkBankDB This can be useful for verifying "corpora" param passed to query functions by walking down and verifying path in object returned here.  Can give user feedback on what part of path is incorrect. Future use for GUIs to select paths or auto-complete
}
