% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{getTokens}
\alias{getTokens}
\title{Get token metadata data}
\usage{
getTokens(
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL
)
}
\arguments{
\item{corpusName}{Name of corpus to query}

\item{corpora}{Name of corpora(s) to query}

\item{lang}{Query by Language}

\item{media}{Query by Media}

\item{age}{Query by Age}

\item{gender}{Query by gender}

\item{designType}{Query by design type}

\item{activityType}{Query by activity type}

\item{groupType}{Query by group type}
}
\description{
Query a table with all the words from the selected transcripts, one word (token) per row. Each row has:
\itemize{
\item Link to view transcript and play any associated media.
\item Corpus path to transcript.
\item Utterance sequence number (starts at 0).
\item Word sequence number within utterance (starts at 0).
\item Speaker's role.
\item Speaker's ID.
\item The word (token).
\item The word's stem.
\item Part of speech code. (See CHAT manual for descriptions of codes).
}
}
\examples{
oneFileTokens <- getTokens(corpusName = list('childes'), corpora = list(list('childes', 'Eng-NA', 'MacWhinney', '010411a')))
}
