% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{getTranscripts}
\alias{getTranscripts}
\title{Get transcript data}
\usage{
getTranscripts(
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL
)
}
\arguments{
\item{corpusName}{Name of corpus to query}

\item{corpora}{Name of corpora(s) to query}

\item{lang}{Query by Language}

\item{media}{Query by Media}

\item{age}{Query by Age}

\item{gender}{Query by gender}

\item{designType}{Query by design type}

\item{activityType}{Query by activity type}

\item{groupType}{Query by group type}
}
\description{
Queries a table where each row represents a transcript matching the query from step 1. Each row has:
\itemize{
\item Link to view transcript and play any associated media.
\item Corpus path to transcript.
\item Media types (audio/video) linked to transcript.
\item Unique ID for transcript (PID).
\item Languages spoken. Date recorded.
\item Design Type.
\item Activity Type.
\item Group Type.
}
}
\examples{
getTranscripts(corpusName = list('childes'),
               corpora = list(list('childes',
                                   'Eng-NA',
                                   'MacWhinney',
                                   '010411a')))
}
