% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_values_functions.R
\name{validPath}
\alias{validPath}
\title{Determine if user's attempted targetPath exists in pathTree.}
\usage{
validPath(targetPath, pathTree = getPathTrees())
}
\arguments{
\item{targetPath}{Vector of strings representing path to validate in pathTree.}

\item{pathTree}{The tree of folder and file paths (value from getPathTrees()).}
}
\description{
This helper function can be used by other functions to validate user 'corpora' parameter values passed to the query functions.
}
\examples{
validPath(c('respMsg', 'childes', 'childes', 'Clinical'));

validPath(c('respMsg', 'childes', 'childes', 'somethingThatDoesNotExist'));
}
