% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{getTokenTypes}
\alias{getTokenTypes}
\title{Get token type metadata}
\usage{
getTokenTypes(
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL
)
}
\arguments{
\item{corpusName}{Name of corpus to query}

\item{corpora}{Name of corpora(s) to query}

\item{lang}{Query by Language}

\item{media}{Query by Media}

\item{age}{Query by Age}

\item{gender}{Query by gender}

\item{designType}{Query by design type}

\item{activityType}{Query by activity type}

\item{groupType}{Query by group type}
}
\description{
Queries a table with all the words from the selected transcripts condensed into "types" based on word form and part of speech. Each row has:
\itemize{
\item Speaker's role.
\item The word. Number of occurances of word in selected transcripts.
\item Part of speech (See CHAT manual for descriptions of codes).
\item The word's stem.
}
}
\examples{
macwTokenTypes <- getTokenTypes(corpusName = list('childes'), corpora = list(list('childes', 'Eng-NA', 'MacWhinney')))
}
