% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validPath}
\alias{validPath}
\title{Check for valid path}
\usage{
validPath(targetPath)
}
\arguments{
\item{targetPath}{Vector of strings representing path to validate in pathTree.}
}
\description{
This helper function can be used by other functions to validate user 'corpora' parameter values passed to the query functions.
}
\examples{
validPath(c('childes', 'childes', 'Clinical'));

validPath(c('childes', 'childes', 'somethingThatDoesNotExist'));
}
